# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "To-Do API",
    "description": "API REST para gerenciamento de to-dos com compartilhamento entre usuários. Construída com Next.js + TypeScript + Supabase.",
    "version": "1.0.0",
    "contact": {
      "name": "API Support"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Servidor de desenvolvimento"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Token JWT obtido no login"
      }
    },
...
  },
  "paths": {
...
    "/api/todos": {
      "get": {
        "summary": "Lista todos os to-dos do usuário",
        "tags": ["Todos"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "in": "query", "name": "page", "schema": { "type": "integer", "default": 1 } },
          { "in": "query", "name": "limit", "schema": { "type": "integer", "default": 10, "maximum": 100 } },
          { "in": "query", "name": "completed", "schema": { "type": "string", "enum": ["true", "false"] } },
          { "in": "query", "name": "priority", "schema": { "type": "string", "enum": ["low", "medium", "high"] } }
        ],
        "responses": {
          "200": { "description": "Lista de to-dos com paginação" },
          "401": { "description": "Não autenticado" }
        }
      },
      "post": {
        "summary": "Cria um novo to-do",
        "tags": ["Todos"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["title"],
                "properties": {
                  "title": { "type": "string", "maxLength": 255 },
                  "description": { "type": "string", "maxLength": 1000 },
                  "priority": { "type": "string", "enum": ["low", "medium", "high"], "default": "medium" },
                  "due_date": { "type": "string", "format": "date-time" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "To-do criado com sucesso" },
          "400": { "description": "Dados inválidos" },
          "401": { "description": "Não autenticado" }
        }
      }
    },
...
    "/api/todos/{id}/share": {
      "get": {
        "summary": "Lista usuários com quem o to-do foi compartilhado",
        "tags": ["Sharing"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "in": "path", "name": "id", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "200": { "description": "Lista de compartilhamentos" },
          "401": { "description": "Não autenticado" },
          "403": { "description": "Sem permissão" },
          "404": { "description": "To-do não encontrado" }
        }
      },
      "post": {
        "summary": "Compartilha um to-do com outro usuário",
        "tags": ["Sharing"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "in": "path", "name": "id", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email"],
                "properties": {
                  "email": { "type": "string", "format": "email" },
                  "permission": { "type": "string", "enum": ["view", "edit"], "default": "view" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Compartilhamento criado" },
          "400": { "description": "Dados inválidos ou já compartilhado" },
          "401": { "description": "Não autenticado" },
          "403": { "description": "Sem permissão" },
          "404": { "description": "To-do ou usuário não encontrado" }
        }
      }
    },
    "/api/todos/{id}/share/{shareId}": {
      "delete": {
        "summary": "Remove um compartilhamento",
        "tags": ["Sharing"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "in": "path", "name": "id", "required": true, "schema": { "type": "string", "format": "uuid" } },
          { "in": "path", "name": "shareId", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "200": { "description": "Compartilhamento removido" },
          "401": { "description": "Não autenticado" },
          "403": { "description": "Sem permissão" },
          "404": { "description": "Compartilhamento não encontrado" }
        }
      },
      "patch": {
        "summary": "Atualiza permissão de um compartilhamento",
        "tags": ["Sharing"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "in": "path", "name": "id", "required": true, "schema": { "type": "string", "format": "uuid" } },
          { "in": "path", "name": "shareId", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["permission"],
                "properties": {
                  "permission": { "type": "string", "enum": ["view", "edit"] }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Permissão atualizada" },
          "401": { "description": "Não autenticado" },
          "403": { "description": "Sem permissão" },
          "404": { "description": "Compartilhamento não encontrado" }
        }
      }
    },
...
  }
}